Further refactoring to reduce circular deps.
