+++ util/chplenv/chpl_arch.py
@@ -9,7 +9,7 @@ chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
 import chpl_comm, chpl_compiler, chpl_platform, overrides
-from etc import CompVersion, compiler_is_prgenv, get_compiler_version
+from compiler_utils import CompVersion, compiler_is_prgenv, get_compiler_version
 from utils import memoize, run_command
