+++ util/chplenv/chpl_comm.py
@@ -6,7 +6,7 @@ chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
 import chpl_compiler, chpl_platform, overrides
-from etc import using_chapel_module
+from chpl_home_utils import using_chapel_module
 from utils import memoize
