+++ util/chplenv/utils.py
@@ -1,34 +1,6 @@
 """ Utility functions for chplenv modules """
 import subprocess
 
-# List of Chapel Environment Variables
-chplvars = [
-             'CHPL_HOME',
-             'CHPL_HOST_PLATFORM',
-             'CHPL_HOST_COMPILER',
-             'CHPL_TARGET_PLATFORM',
-             'CHPL_TARGET_COMPILER',
-             'CHPL_TARGET_ARCH',
-             'CHPL_LOCALE_MODEL',
-             'CHPL_COMM',
-             'CHPL_COMM_SUBSTRATE',
-             'CHPL_GASNET_SEGMENT',
-             'CHPL_TASKS',
-             'CHPL_LAUNCHER',
-             'CHPL_TIMERS',
-             'CHPL_UNWIND',
-             'CHPL_MEM',
-             'CHPL_MAKE',
-             'CHPL_ATOMICS',
-             'CHPL_NETWORK_ATOMICS',
-             'CHPL_GMP',
-             'CHPL_HWLOC',
-             'CHPL_REGEXP',
-             'CHPL_WIDE_POINTERS',
-             'CHPL_LLVM',
-             'CHPL_AUX_FILESYS',
-           ]
-
 
 def memoize(func):
     cache = func.cache = {}
