+++ util/printchplenv
@@ -9,6 +9,13 @@ from chplenv import *
 
 def print_mode(mode='list', anonymize=False):
     print_var.first_time = True
+
+    if mode == 'overrides':
+        for var in overrides.allvars():
+            stdout.write(var);
+            stdout.write('\n');
+        return
+
     # `m` is a 'simplified' version of `mode` that merges some cases that can
     # be treated the same by print_mode, but need to be dealt with separately
     # in print_var
@@ -20,6 +27,7 @@ def print_mode(mode='list', anonymize=False):
         # make and simple are almost the same barring a few cases and when they
         # are printed
         m = 'make'
+
 
     chpl_home = overrides.get('CHPL_HOME', '')
     chpl_module_home = os.environ.get('CHPL_MODULE_HOME', '')
@@ -48,7 +56,7 @@ def print_mode(mode='list', anonymize=False):
         print_var('CHPL_TARGET_PLATFORM', target_platform, mode)
         print_var('CHPL_TARGET_COMPILER', target_compiler, mode)
 
-    get_module_lcd = ( etc.using_chapel_module() and
+    get_module_lcd = ( chpl_home_utils.using_chapel_module() and
                      (mode == 'runtime' or mode == 'make') )
     if m != 'compiler' and m != 'launcher':
         target_arch = chpl_arch.get('target', mode == 'make', get_module_lcd)
@@ -184,11 +192,6 @@ def print_mode(mode='list', anonymize=False):
 
     elif m == 'compiler' or m == 'runtime' or m == 'launcher':
         stdout.write('\n')
-
-    if mode == 'overrides':
-        for var in overrides.allvars():
-            stdout.write(var);
-            stdout.write('\n');
